Revert "[PM-30751] - add secure SSRF protection for internal IPs (#7184)"
This reverts commit 0cba6093049b9945c160a0c5bf3aac1f5d0971bc.
