+++ src/Icons/Util/ServiceCollectionExtension.cs
@@ -2,7 +2,6 @@
 
 using System.Net;
 using AngleSharp.Html.Parser;
-using Bit.Core.Utilities;
 using Bit.Icons.Services;
 
 namespace Bit.Icons.Extensions;
@@ -28,7 +27,7 @@ public static class ServiceCollectionExtension
         {
             AllowAutoRedirect = false,
             AutomaticDecompression = DecompressionMethods.GZip | DecompressionMethods.Deflate,
-        }).AddSsrfProtection(followRedirects: false);
+        });
 
         // The CreatePasswordUri handler wants similar headers as Icons to portray coming from a browser but
         // needs to follow redirects to get the final URL.
@@ -46,7 +45,7 @@ public static class ServiceCollectionExtension
         }).ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler
         {
             AutomaticDecompression = DecompressionMethods.GZip | DecompressionMethods.Deflate,
-        }).AddSsrfProtection();
+        });
     }
 
     public static void AddHtmlParsing(this IServiceCollection services)
